Modifying Code Alarms
issue:https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I7EJSG

+++ ecmascript/regexp/regexp_parser.cpp
@@ -1218,7 +1218,7 @@ int RegExpParser::ParseCharacterEscape()
             break;
         }
     }
-    return result;
+    return static_cast<int>(result);
 }
 
 bool RegExpParser::ParseClassRanges(RangeSet *result)
